refactor(ApplicationTest): mock 과련 함수 파일 분리
